refactor: move issue extraction to jira svc
