+++ gitlab-jira-integration-application/src/main/java/fr/mmarie/core/jira/JiraService.java
@@ -1,5 +1,6 @@
 package fr.mmarie.core.jira;
 
+import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.squareup.okhttp.OkHttpClient;
 import com.squareup.okhttp.Request;
@@ -15,11 +16,19 @@ import java.io.IOException;
 import java.util.Base64;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import static fr.mmarie.utils.Common.sanitizeURL;
 
 @Slf4j
 public class JiraService {
+
+    /**
+     * Matches issues with name like : #TEST-1, does not support special characters in
+     * project name, #TEST-TEST-1 won't match.
+     */
+    public static final Pattern ISSUE_PATTERN = Pattern.compile("#\\s*(\\w+-\\d+)");
 
     private final JiraConfiguration jiraConfiguration;
 
@@ -60,6 +69,24 @@ public class JiraService {
 
     public Response<Map<String, Object>> serverInfo() throws IOException {
         return jiraEndPoints.serverInfo().execute();
+    }
+
+    /**
+     * Extracts issues names from given {@code message}.
+     *
+     * @param message Commit message
+     * @return Matching issues name
+     */
+    public List<String> extractIssuesFromMessage(String message) {
+        List<String> issues = Lists.newArrayList();
+
+        Matcher matcher = ISSUE_PATTERN.matcher(message);
+
+        while (matcher.find()) {
+            issues.add(matcher.group(1));
+        }
+
+        return issues;
     }
 
     public boolean isExistingIssue(String issue) {
