+++ gitlab-jira-integration-application/src/test/java/fr/mmarie/core/jira/JiraServiceTestIT.java
@@ -8,6 +8,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import retrofit.Response;
 
+import java.util.List;
 import java.util.Map;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
@@ -188,5 +189,37 @@ public class JiraServiceTestIT {
 
         wireMockRule.verify(getRequestedFor(urlEqualTo("/rest/api/2/issue/" + issue + "/comment"))
                 .withHeader("Authorization", matching("Basic .*")));
+    }
+
+    @Test
+    public void extractIssuesFromMessageWithoutIssue() throws Exception {
+        String message = "test: no issue";
+
+        final List<String> issues = jiraService.extractIssuesFromMessage(message);
+
+        assertThat(issues)
+                .hasSize(0);
+    }
+
+    @Test
+    public void extractIssuesFromMessageWithOneIssue() throws Exception {
+        String message = "test(#TEST-1): single issue";
+
+        final List<String> issues = jiraService.extractIssuesFromMessage(message);
+
+        assertThat(issues)
+                .hasSize(1)
+                .containsExactly("TEST-1");
+    }
+
+    @Test
+    public void extractIssuesFromMessageWithMoreThanOneIssue() throws Exception {
+        String message = "test(#TEST-1): issue related to #TEST-15289";
+
+        final List<String> issues = jiraService.extractIssuesFromMessage(message);
+
+        assertThat(issues)
+                .hasSize(2)
+                .containsExactly("TEST-1", "TEST-15289");
     }
 }
